Added variable controlled chunk size
Added a chunk size variable and modified pdf_to_text function to support it. Can modify the chunk size in config.py, in addition to this I also added to the system message when uploading documents to get the model to also state where in the document it found the exerpt from however this needs refinement.